report progress from LaunchTask + task weights

+++ launcher/InstanceImportTask.cpp
@@ -202,7 +202,7 @@ void InstanceImportTask::processZipPack()
     });
     connect(zipTask.get(), &Task::stepProgress, this, &InstanceImportTask::propagateStepProgress);
 
-    connect(zipTask.get(), &Task::progress, this, [this, progressStep](qint64 current, qint64 total) {
+    connect(zipTask.get(), &Task::progress, this, [this, progressStep](double current, double total) {
         progressStep->update(current, total);
         stepProgress(*progressStep);
     });
