+++ launcher/ui/dialogs/OfflineLoginDialog.h
@@ -29,7 +29,7 @@ class OfflineLoginDialog : public QDialog {
     void onTaskFailed(const QString& reason);
     void onTaskSucceeded();
     void onTaskStatus(const QString& status);
-    void onTaskProgress(qint64 current, qint64 total);
+    void onTaskProgress(double current, double total);
 
     void on_userTextBox_textEdited(const QString& newText);
     void on_allowLongUsernames_stateChanged(int value);
